Fix KumoCloudAccount unit parsing to consistently handle missing fields (#48)
* Fix KumoCloudAccount unit parsing to consistently handle missing fields (#47)

Previously, the try block in _parse_unit(raw_unit) was outside of the for loop. That meant that as soon as a missing key was encountered, the function stopped iterating through subsequent keys, leading to a partial unit dict, even if more desired fields were present in the original raw_unit.

Additionally, iteration over the 'fields' set is non-deterministic because [Python set iteration is based on a random hash seed](https://stackoverflow.com/questions/3848091/set-iteration-order-varies-from-run-to-run).

These two factors combined could lead to unit parsing that sometimes worked and sometimes didn't. (e.g. if 'address' was missing and iterated through first, a completely empty unit dict would be returned).

* v0.3.8

---------

Co-authored-by: Mike Thomas <mike@mt0321.com>